some first changes
